+++ portfolio_toolkit/__init__.py
@@ -1,1 +1,21 @@
 """Portfolio Toolkit - A comprehensive toolkit for portfolio analysis and management."""
+
+__version__ = "0.1.0"
+__author__ = "Guido Genzone"
+
+# Main imports for easy access
+try:
+    from .portfolio.portfolio import Portfolio
+    from .asset.portfolio_asset import PortfolioAsset
+    from .data_provider.yf_data_provider import YFDataProvider
+except ImportError:
+    # Handle import errors gracefully during development
+    pass
+
+__all__ = [
+    "__version__",
+    "__author__",
+    "Portfolio",
+    "PortfolioAsset", 
+    "YFDataProvider",
+]
