+++ portfolio_toolkit/cli/cli.py
@@ -1,0 +1,32 @@
+import click
+from .commands.clear_cache import clear_cache
+
+# New organized command groups
+from .commands.ticker.ticker import ticker
+from .commands.watchlist.watchlist import watchlist
+from .commands.optimization.optimization import optimization
+from .commands.portfolio.portfolio import portfolio
+
+
+@click.group()
+@click.version_option(version='1.0.0', package_name='portfolio-toolkit')
+def cli():
+    """Portfolio Toolkit CLI - Manage and analyze your investment portfolios."""
+    pass
+
+
+cli.add_command(clear_cache)
+
+# New organized command groups
+cli.add_command(ticker)
+cli.add_command(watchlist)
+cli.add_command(optimization)
+cli.add_command(portfolio)
+
+
+def main():
+    cli()
+
+
+if __name__ == "__main__":
+    main()
