+++ portfolio_toolkit/cli/commands/portfolio/closed_positions.py
