+++ portfolio_toolkit/cli/main.py
@@ -1,0 +1,7 @@
+#!/usr/bin/env python3
+"""Main entry point for the Portfolio Toolkit CLI."""
+
+from .cli import main
+
+if __name__ == "__main__":
+    main()
