+++ pyproject.toml
@@ -1,9 +1,67 @@
+[build-system]
+requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
+build-backend = "setuptools.build_meta"
+
 [project]
-name = "portfolio-toolkit"  
+name = "portfolio-toolkit"
+dynamic = ["version"]
+description = "A comprehensive toolkit for portfolio analysis and management"
+readme = "README.md"
+requires-python = ">=3.9"
+license = {text = "MIT"}
+authors = [
+    {name = "Guido Genzone"},
+]
+keywords = ["portfolio", "finance", "investment", "analysis", "toolkit"]
+dependencies = [
+    "pandas>=2.0.0",
+    "numpy>=1.24.0",
+    "matplotlib>=3.7.0",
+    "yfinance>=0.2.0",
+    "tabulate",
+    "click>=8.0.0",
+]
+classifiers = [
+    "Development Status :: 4 - Beta",
+    "Intended Audience :: Financial and Insurance Industry",
+    "License :: OSI Approved :: MIT License",
+    "Operating System :: OS Independent",
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
+    "Topic :: Office/Business :: Financial :: Investment",
+    "Topic :: Scientific/Engineering :: Information Analysis",
+]
+
+[project.optional-dependencies]
+dev = [
+    "pytest>=7.0.0",
+    "black>=23.0.0",
+    "isort>=5.0.0",
+    "flake8>=6.0.0",
+]
+docs = [
+    "sphinx>=7.0.0",
+    "sphinx-rtd-theme>=1.3.0",
+]
+
+[project.urls]
+Homepage = "https://github.com/ggenzone/portfolio-tools"
+Documentation = "https://ggenzone.github.io/portfolio-tools"
+Repository = "https://github.com/ggenzone/portfolio-tools.git"
+"Bug Tracker" = "https://github.com/ggenzone/portfolio-tools/issues"
+
+[project.scripts]
+portfolio-toolkit = "portfolio_toolkit.cli.main:main"
+
+[tool.setuptools.dynamic]
+version = {attr = "portfolio_toolkit.__version__"}
 
 [tool.black]
 line-length = 88
-target-version = ['py38']
+target-version = ['py39']
 
 [tool.isort]
 profile = "black"
