+++ tests/test_cli_basic.py
@@ -17,7 +17,7 @@ class TestCLIBasic(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
         cls.project_root = Path(__file__).parent.parent
-        cls.cli_module = "cli.cli"
+        cls.cli_module = "portfolio_toolkit.cli.cli"
 
     def run_command(self, args, description):
         """Run a CLI command and check if it succeeds."""
